+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## master
-* Introduce `--cocoapods` option  
+* Introduce `--include-existing` (`--no-include-existing`) option  
   [Toshihiro Suzuki](https://github.com/toshi0383)
-  [#7](https://github.com/toshi0383/xcconfig-extractor/pull/7)
+  [#8](https://github.com/toshi0383/xcconfig-extractor/pull/8)
 
 ## 0.2.0